samples/python/log-scanner-master/requirements.txt
alembic==0.9.6
click==6.7
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
maxminddb==1.3.0
maxminddb-geolite2==2017.803
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.15
Werkzeug==0.12.2
